+++ mitmproxy/test/tflow.py
@@ -26,6 +26,7 @@ def ttcpflow(client_conn=True, server_conn=True, messages=True, err=None) -> tcp
     f = tcp.TCPFlow(client_conn, server_conn)
     f.messages = messages
     f.error = err
+    f.live = True
     return f
 
 
@@ -80,6 +81,7 @@ def twebsocketflow(messages=True, err=None, close_code=None, close_reason='') ->
             # NORMAL_CLOSURE
             flow.websocket.close_code = 1000
 
+    flow.live = True
     return flow
 
 
@@ -91,6 +93,7 @@ def tflow(
     resp: Union[bool, http.Response] = False,
     err: Union[bool, flow.Error] = False,
     ws: Union[bool, websocket.WebSocketData] = False,
+    live: bool = True,
 ) -> http.HTTPFlow:
     """Create a flow for testing."""
     if client_conn is None:
@@ -116,6 +119,7 @@ def tflow(
     f.response = resp or None
     f.error = err or None
     f.websocket = ws or None
+    f.live = live
     return f
 
 
@@ -136,6 +140,7 @@ def tdummyflow(client_conn=True, server_conn=True, err=None) -> DummyFlow:
 
     f = DummyFlow(client_conn, server_conn)
     f.error = err
+    f.live = True
     return f